Move utils.js to utils/parse-template.js.
