+++ src/content/plugins/fastmail.js
@@ -1,7 +1,7 @@
 /* Fastmail plugin
  */
 
-import {parseTemplate} from '../utils.js';
+import parseTemplate from '../utils/parse-template.js';
 import {insertTemplate} from '../editors/editor-universal.js';
 
 var parseList = function (list) {
