+++ src/content/plugins/gmail.js
@@ -1,7 +1,7 @@
 /* Gmail plugin
  */
 
-import {parseTemplate} from '../utils.js'
+import parseTemplate from '../utils/parse-template.js'
 import {insertTemplate} from '../editors/editor-universal.js'
 import {isContentEditable} from '../editors/editor-contenteditable.js'
 import createContact from '../utils/create-contact.js'
