+++ src/content/plugins/outlook.js
@@ -2,7 +2,7 @@
 /* Outlook plugin
  */
 
-import {parseTemplate} from '../utils.js';
+import parseTemplate from '../utils/parse-template.js';
 import {insertTemplate} from '../editors/editor-universal.js';
 import createContact from '../utils/create-contact.js';
 import {enableBubble} from '../bubble.js';
