+++ src/content/plugins/zendesk.js
@@ -1,7 +1,7 @@
 /* Zendesk plugin
  */
 
-import {parseTemplate} from '../utils.js';
+import parseTemplate from '../utils/parse-template.js';
 import {insertTemplate} from '../editors/editor-universal.js';
 import createContact from '../utils/create-contact.js';
