+++ test/test.js
@@ -1,4 +1,4 @@
-import '../src/content/helpers/moment.spec.js';
-import '../src/content/search.spec.js';
-import '../src/content/utils/plain-text.spec.js';
-import '../src/content/utils.spec.js';
+import '../src/content/helpers/moment.spec.js'
+import '../src/content/search.spec.js'
+import '../src/content/utils/plain-text.spec.js'
+import '../src/content/utils/parse-template.spec.js'
